payCalculator

+++ payCalculator.py
@@ -4,8 +4,8 @@ def calculatePay():
     # This first line is provided for you
     hrs = input("Enter Hours: ")
     rate = input("Enter Rate: ")
-    pay = int(hrs) * int(rate)
-    print("Pay:",pay)
+    pay = float(hrs)*float(rate)
+    print("Pay:", pay)
     # end assignment
 
 ## if you want to test locally before you try to sync
